+++ examples/failing-shell-script-test/run_tests.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+# Exit status > 0 signals successful test execution.
+exit 1;