+++ examples/successful-shell-script-test/simple.clj
@@ -1,0 +1,4 @@
+(defci)
+
+(defjob run-tests
+  :test (execute "run_tests.sh"))