+++ simple.clj
@@ -1,0 +1,7 @@
+(defci)
+
+(defjob test-simple-ci
+  :test (execute "run-tests.sh")
+  :triggers [{:type :timer
+              :name "test-simple-ci-trigger"
+              :args [:every 5 :minutes]}])